+++ framework-crates/objc2-core-video/src/lib.rs
@@ -1,0 +1,25 @@
+//! # Bindings to the `CoreVideo` framework
+//!
+//! See [Apple's docs][apple-doc] and [the general docs on framework crates][framework-crates] for more information.
+//!
+//! [apple-doc]: https://developer.apple.com/documentation/corevideo/
+//! [framework-crates]: https://docs.rs/objc2/latest/objc2/topics/about_generated/index.html
+#![no_std]
+#![cfg_attr(docsrs, feature(doc_auto_cfg))]
+// Update in Cargo.toml as well.
+#![doc(html_root_url = "https://docs.rs/objc2-core-video/0.2.2")]
+
+#[cfg(feature = "alloc")]
+extern crate alloc;
+
+#[cfg(feature = "std")]
+extern crate std;
+
+mod generated;
+#[allow(unused_imports, unreachable_pub)]
+pub use self::generated::*;
+
+#[allow(dead_code)]
+pub(crate) type Boolean = u8;
+#[allow(dead_code)]
+pub(crate) type OSType = u32;