+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Added new framework crates:
   - `CoreAudioTypes` / `objc2-core-audio-types`.
   - `CoreFoundation` / `objc2-core-foundation`.
+  - `CoreVideo` / `objc2-core-video`.
   - `ScreenSaver` / `objc2-screen-saver`.
 
 ### Changed